ProdutoController refatorada, funcionando! bug resolvido.

+++ src/main/java/com/bootcamp_w3_g3/controller/ProdutoController.java
@@ -63,7 +63,7 @@ public class ProdutoController {
      * @return produtoDTO
      */
     @DeleteMapping(value="/deletar/{cod}")
-    public Produto apagar(@PathVariable Integer cod)
+    public Produto apagar(@PathVariable Long cod)
     {
         return produtoService.apagar(cod);
     }
